api/requirements.txt
Flask==2.3.3
google-cloud-pubsub==2.21.0
